% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Perform PCA on dataframe}
\usage{
do_pca(.data, product, liking = NULL, scale = TRUE)
}
\arguments{
\item{.data}{a dataframe or tibble.}

\item{product}{name of column containing product information.}

\item{liking}{name of column containing hedonic liking (default to NULL).}

\item{scale}{whether to scale data to unit of variance (default to TRUE)}
}
\value{
Returns a list containing information of PCA results.
}
\description{
Perform PCA on dataframe
}
\examples{
sensochoc_means \%>\% do_pca(product = "Product")
do_pca(sensochoc_means, product = "Product")

}
